esm: refactor test-esm-loader-resolve-type
